+++ tests/Phormium/Tests/FilterTest.php
@@ -5,7 +5,7 @@ namespace Phormium\Tests;
 use \Phormium\Tests\Models\Person;
 
 use \Phormium\DB;
-use \Phormium\Filter;
+use \Phormium\ColumnFilter;
 use \Phormium\Parser;
 
 class FilterTest extends \PHPUnit_Framework_TestCase
@@ -17,7 +17,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
 
     public function testEq()
     {
-        $filter = new Filter('test', '=', 1);
+        $filter = new ColumnFilter('test', '=', 1);
         $actual = $filter->render();
         $expected = array("test = ?", array(1));
         self::assertSame($expected, $actual);
@@ -25,7 +25,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
 
     public function testNeq1()
     {
-        $filter = new Filter('test', '!=', 1);
+        $filter = new ColumnFilter('test', '!=', 1);
         $actual = $filter->render();
         $expected = array("test != ?", array(1));
         self::assertSame($expected, $actual);
@@ -33,7 +33,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
 
     public function testNeq2()
     {
-        $filter = new Filter('test', '<>', 1);
+        $filter = new ColumnFilter('test', '<>', 1);
         $actual = $filter->render();
         $expected = array("test <> ?", array(1));
         self::assertSame($expected, $actual);
@@ -41,7 +41,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
 
     public function testGt()
     {
-        $filter = new Filter('test', '>', 1);
+        $filter = new ColumnFilter('test', '>', 1);
         $actual = $filter->render();
         $expected = array("test > ?", array(1));
         self::assertSame($expected, $actual);
@@ -49,7 +49,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
 
     public function testGte()
     {
-        $filter = new Filter('test', '>=', 1);
+        $filter = new ColumnFilter('test', '>=', 1);
         $actual = $filter->render();
         $expected = array("test >= ?", array(1));
         self::assertSame($expected, $actual);
@@ -57,7 +57,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
 
     public function testLt()
     {
-        $filter = new Filter('test', '<', 1);
+        $filter = new ColumnFilter('test', '<', 1);
         $actual = $filter->render();
         $expected = array("test < ?", array(1));
         self::assertSame($expected, $actual);
@@ -65,7 +65,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
 
     public function testLte()
     {
-        $filter = new Filter('test', '<=', 1);
+        $filter = new ColumnFilter('test', '<=', 1);
         $actual = $filter->render();
         $expected = array("test <= ?", array(1));
         self::assertSame($expected, $actual);
@@ -73,7 +73,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
 
     public function testIn()
     {
-        $filter = new Filter('test', 'in', array(1, 2, 3));
+        $filter = new ColumnFilter('test', 'in', array(1, 2, 3));
         $actual = $filter->render();
         $expected = array("test IN (?, ?, ?)", array(1, 2, 3));
         self::assertSame($expected, $actual);
@@ -85,13 +85,13 @@ class FilterTest extends \PHPUnit_Framework_TestCase
      */
     public function testInWrongParam()
     {
-        $filter = new Filter('test', 'in', 1);
+        $filter = new ColumnFilter('test', 'in', 1);
         $filter->render();
     }
 
     public function testNotIn()
     {
-        $filter = new Filter('test', 'not in', array(1, 2, 3));
+        $filter = new ColumnFilter('test', 'not in', array(1, 2, 3));
         $actual = $filter->render();
         $expected = array("test NOT IN (?, ?, ?)", array(1, 2, 3));
         self::assertSame($expected, $actual);
@@ -103,13 +103,13 @@ class FilterTest extends \PHPUnit_Framework_TestCase
      */
     public function testNotInWrongParam()
     {
-        $filter = new Filter('test', 'not in', 1);
+        $filter = new ColumnFilter('test', 'not in', 1);
         $filter->render();
     }
 
     public function testIsNull()
     {
-        $filter = new Filter('test', 'is null');
+        $filter = new ColumnFilter('test', 'is null');
         $actual = $filter->render();
         $expected = array("test IS NULL", array());
         self::assertSame($expected, $actual);
@@ -117,7 +117,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
 
     public function testNotNull()
     {
-        $filter = new Filter('test', 'not null');
+        $filter = new ColumnFilter('test', 'not null');
         $actual = $filter->render();
         $expected = array("test IS NOT NULL", array());
         self::assertSame($expected, $actual);
@@ -125,7 +125,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
 
     public function testLike()
     {
-        $filter = new Filter('test', 'like', '%foo%');
+        $filter = new ColumnFilter('test', 'like', '%foo%');
         $actual = $filter->render();
         $expected = array("test LIKE ?", array('%foo%'));
         self::assertSame($expected, $actual);
@@ -133,7 +133,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
 
     public function testNotLike()
     {
-        $filter = new Filter('test', 'not like', '%bar%');
+        $filter = new ColumnFilter('test', 'not like', '%bar%');
         $actual = $filter->render();
         $expected = array("test NOT LIKE ?", array('%bar%'));
         self::assertSame($expected, $actual);
@@ -141,7 +141,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
 
     public function testBetween()
     {
-        $filter = new Filter('test', 'between', array(10, 20));
+        $filter = new ColumnFilter('test', 'between', array(10, 20));
         $actual = $filter->render();
         $expected = array("test BETWEEN ? AND ?", array(10, 20));
         self::assertSame($expected, $actual);
@@ -153,7 +153,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
      */
     public function testBetweenWrongParam1()
     {
-        $filter = new Filter('test', 'between', 'xxx');
+        $filter = new ColumnFilter('test', 'between', 'xxx');
         $filter->render();
     }
 
@@ -163,7 +163,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
      */
     public function testBetweenWrongParam2()
     {
-        $filter = new Filter('test', 'between', array(1));
+        $filter = new ColumnFilter('test', 'between', array(1));
         $filter->render();
     }
 
@@ -173,7 +173,7 @@ class FilterTest extends \PHPUnit_Framework_TestCase
      */
     public function testUnknownOp()
     {
-        $filter = new Filter('test', 'xxx');
+        $filter = new ColumnFilter('test', 'xxx');
         $filter->render();
     }
 
@@ -191,5 +191,48 @@ class FilterTest extends \PHPUnit_Framework_TestCase
 
         self::assertSame(4, $qs->count());
         self::assertCount(4, $qs->fetch());
+    }
+
+    public function testFilterFromArray()
+    {
+        $actual = ColumnFilter::fromArray(['id', '=', 123]);
+
+        self::assertInstanceOf('\\Phormium\\ColumnFilter', $actual);
+        self::assertSame('id', $actual->column);
+        self::assertSame('=', $actual->operation);
+        self::assertSame(123, $actual->value);
+
+        $actual = ColumnFilter::fromArray(['email', 'null']);
+
+        self::assertInstanceOf('\\Phormium\\ColumnFilter', $actual);
+        self::assertSame('email', $actual->column);
+        self::assertSame('NULL', $actual->operation);
+        self::assertNull($actual->value);
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage Invalid filter sepecification
+     */
+    public function testFilterFromArrayExceptionTooMany()
+    {
+        $actual = ColumnFilter::fromArray([1, 2, 3, 4, 5]);
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage Invalid filter sepecification
+     */
+    public function testFilterFromArrayExceptionTooFew()
+    {
+        $actual = ColumnFilter::fromArray([1]);
+    }
+
+    /**
+     * @expectedException \Exception
+     */
+    public function testFilterFromArrayExceptionWrongType()
+    {
+        $actual = ColumnFilter::fromArray(1);
     }
 }