+++ CHANGELOG.md
@@ -4,8 +4,10 @@ Phormium Changelog
 0.2 / TBA
 ---------
 
-* Implemented transactions
+* Added transactions
 * Added QuerySet::dump()
+* Added logging via [Apache log4php](http://logging.apache.org/log4php/)
+* Added composite filters
 
 0.1 / 2013-04-25
 ----------------